Huge update
Added epubcheck to the repository so that compiling is as easy as a "python3 CreateE-book.py && java -jar epubcheck.jar MakingGames.epub" command. Started Chapter 10 and added anchor tags for all started chapters. Still need to test on Win10 to see if it compiles in CMD or PowerShell, or requires Bash for Windows 10.